Remove gen-graphs dependency from sample
