+++ gen-core/Cargo.toml
@@ -16,6 +16,7 @@ gen-capnp-schemas = { path = "../gen-capnp-schemas", version = "0.2.1" }
 itertools = "0.14.0"
 capnp = { version = "0.24" }
 noodles = { version = "0.101.0", features = ["gff"] }
+petgraph = "0.6.5"
 rusqlite = { package = "rusqdoltlite", version = "0.40.14", features = ["bundled", "fallible_uint"] }
 serde = {  version = "1.0.219", features = ["derive"] }
 thiserror = "1.0"
